:recycle: make zaak based permissions DRYer
